requirements.txt
apiclient==1.0.4
cachetools==3.1.1
certifi==2022.9.24
charset-normalizer==2.1.1
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
gspread==5.6.0
httplib2==0.20.4
idna==3.4
numpy==1.23.4
oauth2client==4.1.3
oauthlib==3.2.1
protobuf==4.21.7
pyasn1==0.4.6
pyasn1-modules==0.2.6
pyparsing==3.0.9
pyTelegramBotAPI==4.7.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.0
six==1.12.0
telebot==0.0.4
uritemplate==4.1.1
urllib3==1.26.12
